Remove md5 checks.
I wanted to use the archive's md5 to avoid uploading redundant backups,
but to be useful I'd need to remove the random nonce which reduces
security. In practice the schedule should reflect how often the
underlying data changes. It's also of limited utility since only one bit
of one file would need to change to trigger a new backup.